Rename mod scroll_traits -> autoimpl_traits
