+++ crates/kas-macros/src/autoimpl_traits.rs
